# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "failureCode": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "failureCode",
          "message"
        ],
        "type": "object"
      },
...
      "WagerOutcomeResponse": {
        "additionalProperties": false,
        "properties": {
          "balance": {
            "additionalProperties": false,
            "properties": {
              "amount": {
                "pattern": "^\\d+\\.\\d{2}$",
                "type": "string"
              },
              "currency": {
                "pattern": "^[A-Z]{3}$",
                "type": "string"
              }
            },
            "required": [
              "amount",
              "currency"
            ],
            "type": "object"
          },
          "failureCode": {
            "type": "string"
          },
          "idempotentReplay": {
            "type": "boolean"
          },
          "status": {
            "enum": [
              "FAILED",
              "PENDING_REFERENCE",
              "PROCESSED",
              "REJECTED"
            ],
            "type": "string"
          },
          "transactionId": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          }
        },
        "required": [
          "transactionId",
          "status",
          "balance",
          "idempotentReplay"
        ],
        "type": "object"
      },
      "WagerRequest": {
        "additionalProperties": false,
        "properties": {
          "externalTransactionId": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "gameId": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "kind": {
            "enum": [
              "BET",
              "WIN",
              "LOSS",
              "REFUND",
              "ROLLBACK"
            ],
            "type": "string"
          },
          "money": {
            "additionalProperties": false,
            "properties": {
              "amount": {
                "pattern": "^\\d+\\.\\d{2}$",
                "type": "string"
              },
              "currency": {
                "pattern": "^[A-Z]{3}$",
                "type": "string"
              }
            },
            "required": [
              "amount",
              "currency"
            ],
            "type": "object"
          },
          "playerId": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "providerId": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "referenceExternalTransactionId": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "roundId": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "walletId": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          }
        },
        "required": [
          "providerId",
          "externalTransactionId",
          "playerId",
          "walletId",
          "roundId",
          "gameId",
          "kind",
          "money"
        ],
        "type": "object"
      },
      "WagerTransactionResponse": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$",
            "type": "string"
          },
          "externalTransactionId": {
            "type": "string"
          },
          "failureCode": {
            "type": [
              "string",
              "null"
            ]
          },
          "gameId": {
            "type": "string"
          },
          "idempotencyKey": {
            "type": "string"
          },
          "kind": {
            "enum": [
              "BET",
              "WIN",
              "LOSS",
              "REFUND",
              "ROLLBACK"
            ],
            "type": "string"
          },
          "money": {
            "additionalProperties": false,
            "properties": {
              "amount": {
                "pattern": "^\\d+\\.\\d{2}$",
                "type": "string"
              },
              "currency": {
                "pattern": "^[A-Z]{3}$",
                "type": "string"
              }
            },
            "required": [
              "amount",
              "currency"
            ],
            "type": "object"
          },
          "observedBalance": {
            "anyOf": [
              {
                "additionalProperties": false,
                "properties": {
                  "amount": {
                    "pattern": "^\\d+\\.\\d{2}$",
                    "type": "string"
                  },
                  "currency": {
                    "pattern": "^[A-Z]{3}$",
                    "type": "string"
                  }
                },
                "required": [
                  "amount",
                  "currency"
                ],
                "type": "object"
              },
              {
                "type": "null"
              }
            ]
          },
          "playerId": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "processedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "providerId": {
            "type": "string"
          },
          "referenceExternalTransactionId": {
            "type": [
              "string",
              "null"
            ]
          },
          "referenceTransactionId": {
            "anyOf": [
              {
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "roundId": {
            "type": "string"
          },
          "status": {
            "enum": [
              "FAILED",
              "PENDING_REFERENCE",
              "PROCESSED",
              "REJECTED"
            ],
            "type": "string"
          },
          "transactionId": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "walletId": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          }
        },
        "required": [
          "transactionId",
          "providerId",
          "externalTransactionId",
          "idempotencyKey",
          "walletId",
          "playerId",
          "roundId",
          "gameId",
          "kind",
          "money",
          "referenceExternalTransactionId",
          "referenceTransactionId",
          "status",
          "failureCode",
          "observedBalance",
          "processedAt",
          "createdAt"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "HTTP contract for wallet, wagering, reconciliation, and operational endpoints.",
    "title": "Distributed Wagering Processor API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/providers/{providerId}/wagering/transactions/{externalTransactionId}": {
      "get": {
        "operationId": "getProviderWagerTransaction",
        "parameters": [
          {
            "description": "Provider identity",
            "in": "path",
            "name": "providerId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "External transaction identity",
            "in": "path",
            "name": "externalTransactionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "createdAt": "2026-01-01T00:00:00.000Z",
                  "externalTransactionId": "external-transaction-1",
                  "failureCode": null,
                  "gameId": "game-1",
                  "idempotencyKey": "idempotency-key-1",
                  "kind": "BET",
                  "money": {
                    "amount": "20.00",
                    "currency": "BRL"
                  },
                  "observedBalance": {
                    "amount": "80.00",
                    "currency": "BRL"
                  },
                  "playerId": "22222222-2222-4222-8222-222222222222",
                  "processedAt": "2026-01-01T00:00:01.000Z",
                  "providerId": "provider-a",
                  "referenceExternalTransactionId": null,
                  "referenceTransactionId": null,
                  "roundId": "round-1",
                  "status": "PROCESSED",
                  "transactionId": "33333333-3333-4333-8333-333333333333",
                  "walletId": "11111111-1111-4111-8111-111111111111"
                },
                "schema": {
                  "$ref": "#/components/schemas/WagerTransactionResponse"
                }
              }
            },
            "description": "Transaction found"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "failureCode": "INVALID_PAYLOAD",
                  "message": "Request payload is invalid"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Transaction not found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Dependency temporarily unavailable"
          }
        },
        "summary": "Get a wagering transaction by provider identity",
        "tags": [
          "Wagering"
        ]
      }
    },
    "/wagering/transactions": {
      "post": {
        "operationId": "processWagerTransaction",
        "parameters": [
          {
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-correlation-id",
            "required": false,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "externalTransactionId": "external-transaction-1",
                "gameId": "game-1",
                "kind": "BET",
                "money": {
                  "amount": "20.00",
                  "currency": "BRL"
                },
                "playerId": "22222222-2222-4222-8222-222222222222",
                "providerId": "provider-a",
                "roundId": "round-1",
                "walletId": "11111111-1111-4111-8111-111111111111"
              },
              "schema": {
                "$ref": "#/components/schemas/WagerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "example": {
                  "balance": {
                    "amount": "80.00",
                    "currency": "BRL"
                  },
                  "idempotentReplay": false,
                  "status": "PROCESSED",
                  "transactionId": "33333333-3333-4333-8333-333333333333"
                },
                "schema": {
                  "$ref": "#/components/schemas/WagerOutcomeResponse"
                }
              }
            },
            "description": "Transaction processed"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WagerOutcomeResponse"
                }
              }
            },
            "description": "Pending reference accepted"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "failureCode": "INVALID_PAYLOAD",
                  "message": "Request payload is invalid"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Idempotency conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WagerOutcomeResponse"
                }
              }
            },
            "description": "Business transaction rejected"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Dependency temporarily unavailable"
          }
        },
        "summary": "Process an idempotent wagering transaction",
        "tags": [
          "Wagering"
        ]
      }
    },
    "/wagering/transactions/{transactionId}": {
      "get": {
        "operationId": "getWagerTransaction",
        "parameters": [
          {
            "description": "Transaction UUID",
            "in": "path",
            "name": "transactionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "createdAt": "2026-01-01T00:00:00.000Z",
                  "externalTransactionId": "external-transaction-1",
                  "failureCode": null,
                  "gameId": "game-1",
                  "idempotencyKey": "idempotency-key-1",
                  "kind": "BET",
                  "money": {
                    "amount": "20.00",
                    "currency": "BRL"
                  },
                  "observedBalance": {
                    "amount": "80.00",
                    "currency": "BRL"
                  },
                  "playerId": "22222222-2222-4222-8222-222222222222",
                  "processedAt": "2026-01-01T00:00:01.000Z",
                  "providerId": "provider-a",
                  "referenceExternalTransactionId": null,
                  "referenceTransactionId": null,
                  "roundId": "round-1",
                  "status": "PROCESSED",
                  "transactionId": "33333333-3333-4333-8333-333333333333",
                  "walletId": "11111111-1111-4111-8111-111111111111"
                },
                "schema": {
                  "$ref": "#/components/schemas/WagerTransactionResponse"
                }
              }
            },
            "description": "Transaction found"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "failureCode": "INVALID_PAYLOAD",
                  "message": "Request payload is invalid"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Transaction not found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Dependency temporarily unavailable"
          }
        },
        "summary": "Get a wagering transaction by internal identity",
        "tags": [
          "Wagering"
        ]
      }
    },
...
  },
...
}